Allow to write in a buffer depending on the PID wich the app use and which screen was active when the process with a non linked pid was writing the message.
After the link the proccess will be automatically redirected through the buffer_Manager to write in the correct buffer and if the active screen is corresponding with the writing bufffer all will be displayed on the screen.

The buffer will act exactly the same as the screen.

When you witch the screnn by typing ctrl + F1 to F12 all the buffer content will be copyed in the video memory to display it.